+++ modules/libraries/queue.py
@@ -6,8 +6,8 @@ Copyright (c) 2025 TAPS OSS
 Project: https://github.com/TAPSOSS/Walrio
 Licensed under the BSD-3-Clause License (see LICENSE file for details)
 
-A script that manages a queue of songs from the audio library and plays them using audio_player.py and a db file (assumed to be generated by audio_library.py).
-Sample Usage: python audio_queue.py [--db-path <database_path>] [--shuffle] [--repeat] [--artist <artist>] [--album <album>] [--genre <genre>]
+A script that manages a queue of songs from the audio library and plays them using player.py and a db file (assumed to be generated by database.py).
+Sample Usage: python queue.py [--db-path <database_path>] [--shuffle] [--repeat] [--artist <artist>] [--album <album>] [--genre <genre>]
 """
 
 import sys
@@ -19,16 +19,16 @@ import subprocess
 import hashlib
 import time
 from pathlib import Path
-from audio_player import play_audio
+from player import play_audio
 from playlist import load_m3u_playlist
 
-# Try to import audio_library functions for auto-adding missing songs
+# Try to import database functions for auto-adding missing songs
 try:
-    from audio_library import extract_metadata, get_file_hash
-    AUDIO_LIBRARY_AVAILABLE = True
+    from database import extract_metadata, get_file_hash
+    DATABASE_AVAILABLE = True
 except ImportError:
-    print("Note: audio_library.py functions not available. Auto-adding missing songs disabled.")
-    AUDIO_LIBRARY_AVAILABLE = False
+    print("Note: database.py functions not available. Auto-adding missing songs disabled.")
+    DATABASE_AVAILABLE = False
 
 # Default database path
 DEFAULT_DB_PATH = "walrio_library.db"
@@ -37,7 +37,7 @@ DEFAULT_DB_PATH = "walrio_library.db"
 def connect_to_database(db_path):
     if not os.path.exists(db_path):
         print(f"Error: Database file '{db_path}' not found.")
-        print("Please run audio_library.py first to create the database.")
+        print("Please run database.py first to create the database.")
         return None
     try:
         conn = sqlite3.connect(db_path)
@@ -166,7 +166,7 @@ def play_queue(queue, shuffle=False, repeat=False, start_index=0, conn=None):
                 continue
             
             # Auto-add missing song to database if connection is available
-            if conn and AUDIO_LIBRARY_AVAILABLE:
+            if conn and DATABASE_AVAILABLE:
                 try:
                     add_missing_song_to_database(file_path, conn)
                 except Exception as e:
@@ -206,7 +206,7 @@ def play_queue(queue, shuffle=False, repeat=False, start_index=0, conn=None):
 
 # Add missing song to database
 def add_missing_song_to_database(file_path, conn):
-    if not AUDIO_LIBRARY_AVAILABLE:
+    if not DATABASE_AVAILABLE:
         return False
     
     try:
@@ -386,7 +386,7 @@ def interactive_mode(conn):
 def main():
     parser = argparse.ArgumentParser(
         description="Audio Queue Manager - Play songs from your music library",
-        epilog="Example: python audio_queue.py --artist 'Pink Floyd' --shuffle OR python audio_queue.py --playlist myplaylist.m3u"
+        epilog="Example: python queue.py --artist 'Pink Floyd' --shuffle OR python queue.py --playlist myplaylist.m3u"
     )
     parser.add_argument(
         "--db-path",
@@ -447,7 +447,7 @@ def main():
         
         # Try to connect to database for auto-adding missing songs
         conn = None
-        if AUDIO_LIBRARY_AVAILABLE:
+        if DATABASE_AVAILABLE:
             try:
                 if os.path.exists(args.db_path):
                     conn = sqlite3.connect(args.db_path)
